% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\name{hdacCCR_config}
\alias{hdacCCR_config}
\title{The configuration table to analyze \link{hdacCCR_data}.}
\description{
The configuration table to analyze \link{hdacCCR_data}.
}
\details{
\code{hdacCCR_config} is a data frame that specifies the experiment
 names, isobaric labels, and the administered drug concentrations at each
 label.
}
