% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\name{hdacCCR_config_repl1}
\alias{hdacCCR_config_repl1}
\title{The configuration table to analyze \link{hdacCCR_data_repl1}.}
\description{
The configuration table to analyze \link{hdacCCR_data_repl1}.
}
\details{
\code{hdacCCR_config_repl1} is a data frame that specifies the experiment
 name, isobaric labels, and the administered drug concentrations at each
 label.
}

