% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\name{hdacCCR_config_repl2}
\alias{hdacCCR_config_repl2}
\title{The configuration table to analyze \link{hdacCCR_data_repl2}.}
\description{
The configuration table to analyze \link{hdacCCR_data_repl2}.
}
\details{


\code{hdacCCR_config_repl2} is a data frame that specifies the experiment
 name, isobaric labels, and the administered drug concentrations at each
 label.
}

