% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\name{hdacCCR_data_repl2}
\alias{hdacCCR_data_repl2}
\title{TPP-CCR example dataset (replicate 2)}
\description{
Example subset of a Panobinostat TPP-CCR dataset (replicate 2)
}
\details{


A subset of a dataset obtained by TPP-CCR experiments to investigate
 drug effects for HDAC inhibitor Panobinostat. It contains 7 HDACs as well as
 a random selection of 493 further proteins.

 You can use this dataset to explore the \code{\link{TPP}} package
 functionalities without invoking the whole time consuming analysis on the
 big dataset.

 The originial dataset is located in the folder
 \code{'example_data/CCR_example_data'} in the package's installation
 directory. You can find it on your system by the \code{R} command
 \code{system.file('example_data', package = 'TPP')}.
}

