% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\name{hdacTR_config}
\alias{hdacTR_config}
\title{The configuration table to analyze \link{hdacTR_data}.}
\description{
The configuration table to analyze \link{hdacTR_data}.
}
\details{
\code{hdacTR_config} is a data frame that specifies the experiment 
 name, isobaric labels, and the administered temperatures at each 
 label.
}
\references{
Franken, H, Mathieson, T, Childs, D. Sweetman, G. Werner, T. Huber, W. & Savitski, M. M. (2015),
  Thermal proteome profiling for unbiased identification of drug targets and detection of downstream effectors.
  Nature protocols 10(10), 1567-1593.
}
\seealso{
\code{\link{hdacTR_smallExample}}, \code{\link{hdacTR_data}}
}
