% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacTR_fittedData_smallExample.R
\docType{data}
\name{hdacTR_fittedData_smallExample}
\alias{hdacTR_fittedData_smallExample}
\title{Example of a TPP-TR dataset with fitted melting curve parameters.}
\description{
Example of a TPP-TR dataset with fitted melting curve parameters.
}
\details{
Contains the data object \code{\link{trDataFitted}}.
}

