% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacTR_resultsTable_smallExample.R
\docType{data}
\name{hdacTR_resultsTable_smallExample}
\alias{hdacTR_resultsTable_smallExample}
\title{Example of a TPP-TR result table.}
\description{
Example of a TPP-TR result table.
}
\details{
Contains the data object \code{\link{resultTable}}.
}

