% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\docType{data}
\name{hdacTR_smallExample}
\alias{hdacTR_smallExample}
\title{Example subset of a Panobinostat TPP-TR dataset and the
  corresponding configuration table to start the analysis.}
\description{
Example dataset obtained by TPP-TR experiments for analysis by 
the TPP-package. It contains all necessary arguments to start the analysis
(config table and list of data frames).
}

