% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HepG2_trRefData.R
\docType{data}
\name{HepG2_trRefData}
\alias{HepG2_trRefData}
\title{Example of a reference dataset for 2D-TPP experiments.}
\description{
Reference dataset obtained by TPP-TR experiments without drug
treatment on HepG2 cell lines.
}
\details{
Contains the data object \code{\link{tppRefData}}.
}

