% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panobinostat_2DTPP_smallExample.R
\name{panobinostat_2DTPP_config}
\alias{panobinostat_2DTPP_config}
\title{The configuration table to analyze \link{panobinostat_2DTPP_data}.}
\description{
The configuration table to analyze \link{panobinostat_2DTPP_data}.
}
\details{
\code{panobinostat_2DTPP_config} is a data frame that specifies the experiment
 names, isobaric labels, and the administered drug concentrations at each
 label.
}

