% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacTR_resultsTable_smallExample.R
\name{resultTable}
\alias{resultTable}
\title{Example of a TPP-TR result table.}
\description{
Example of a TPP-TR result table.
}
\details{


\code{resultTable} is a data frame that contains the measurements of
 several TPP-TR experiments, the fitted melting curve parameters, as well as
 p-values and the results of additional quality checks for each protein. It
 can be used as input for the function
 \code{\link{tppQCPlotsCorrelateExperiments}}.
}

