% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tppDefunct.R
\name{TPP-defunct}
\alias{TPP-defunct}
\alias{tpp2dPlotCCRGoodCurves}
\alias{tpp2dPlotCCRAllCurves}
\alias{tpp2dPlotCCRSingleCurves}
\alias{tpp2dEvalConfigTable}
\alias{tpp2dRemoveZeroSias}
\alias{tpp2dReplaceColNames}
\alias{tpp2dCreateCCRConfigFile}
\title{Defunct functions in package \sQuote{TPP}}
\usage{
tpp2dPlotCCRGoodCurves()

tpp2dPlotCCRAllCurves()

tpp2dPlotCCRSingleCurves()

tpp2dEvalConfigTable()

tpp2dRemoveZeroSias()

tpp2dReplaceColNames()

tpp2dCreateCCRConfigFile()
}
\value{
No value returned
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \code{tpp2dPlotCCRGoodCurves}, \code{tpp2dPlotCCRAllCurves}, 
\code{tpp2dPlotCCRSingleCurves}, \code{tpp2dEvalConfigTable}, \code{tpp2dRemoveZeroSias},
\code{tpp2dReplaceColNames}, \code{tpp2dCreateCCRConfigFile}
}
