% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tppDeprecated.R
\name{TPP-deprecated}
\alias{TPP-deprecated}
\title{Deprecated functions in package \sQuote{TPP}}
\value{
No value returned
}
\description{
These functions are deprecated and no longer available.
 
%@details The following functions are deprecated; use
%  the replacement indicated below:
%  \itemize{
%  \item
%  }
}
