% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPP.R
\name{TPP}
\alias{TPP-package}
\title{Thermal proteome profiling (TPP)}
\arguments{
\item{libname}{a character string giving the library directory where the 
package defining the namespace was found. Passed to .onLoad function.}

\item{pkgname}{a character string giving the name of the package. Passed to 
.onLoad function.}
}
\value{
No return value defined for this document.
}
\description{
\cite{TPP} is a toolbox for analyzing thermal proteome profiling
  (TPP) experiments.
}
\details{
In order to start a TPP-TR analysis, use function 
  \code{\link{analyzeTPPTR}}. For a TPP-CCR analysis, use function 
  \code{\link{analyzeTPPCCR}}. See the vignette for detailed instructions.
}
\references{
Savitski, M. M., Reinhard, F. B., Franken, H., Werner, T., 
  Savitski, M. F., Eberhard, D., ... & Drewes, G. (2014). Tracking cancer 
  drugs in living cells by thermal profiling of the proteome. Science, 
  346(6205), 1255784.
  
  Franken, H, Mathieson, T, Childs, D. Sweetman, G. Werner, T. Huber, W. & Savitski, M. M. (2015),
  Thermal proteome profiling for unbiased identification of drug targets and detection of downstream effectors.
  Nature protocols 10(10), 1567-1593.
}
