% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dComputeFoldChanges.R
\name{tpp2dComputeFoldChanges}
\alias{tpp2dComputeFoldChanges}
\title{Compute 2D-TPP fold changes}
\usage{
tpp2dComputeFoldChanges(configTable = NULL, data = NULL,
  intensityStr = "signal_sum_", fcStr = "rel_fc_")
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{data}{dataframe that contain the data for the 2D-TPP experiment}

\item{intensityStr}{character string indicating which columns contain the actual 
sumionarea values. Those column names containing the suffix \code{intensityStr} 
will be regarded as containing sumionarea values.}

\item{fcStr}{character string indicating how columns that will contain the actual 
fold change values will be called. The suffix \code{fcStr} will be pasted in front of
the names of the experiments.}
}
\value{
A data.frame with additional columns with constitute fold changes calculated with 
 respect to the intensity values of the zero treatment column
}
\description{
Computes fold changes by calculating fold changes of the sumionarea 
 relative to the reference column.
}
\examples{
data("panob2D_isobQuant_example")
cfg <- panobinostat_2DTPP_config
datRaw <- panobinostat_2DTPP_data
data2d <- tpp2dImport(cfg, datRaw, fcStr = NULL)
fcData2d <- tpp2dComputeFoldChanges(cfg, data2d)
                                 
}

