% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dCreateCCRConfigFile.R
\name{tpp2dCreateCCRConfigFile}
\alias{tpp2dCreateCCRConfigFile}
\title{Create TPP-CCR config files for 2D-TPP experiment}
\usage{
tpp2dCreateCCRConfigFile(configTable)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}
}
\value{
A config file of type data.frame which can be used for the tpp2dRunTPPCCR function
}
\description{
Creates a list of config files for the TPP-CCR analysis so that the config files in the 
 list matches the data.list generated by the 2D-TPP workflow.
}
\examples{
  data("panobinostat_2DTPP_smallExample")         
  CCRconfig <- tpp2dCreateCCRConfigFile(configTable = panobinostat_2DTPP_config)           
}

