% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dCreateDataFrameList.R
\name{tpp2dCreateDataFrameList}
\alias{tpp2dCreateDataFrameList}
\title{Create data frame list for 2D-TPP experiment}
\usage{
tpp2dCreateDataFrameList(configTable = NULL, data = NULL,
  idVar = "representative", fcStr = NULL, addCol = NULL,
  intensityStr = "sumionarea_protein_", qualColName = c("qupm", "qusm"))
}
\arguments{
\item{configTable}{data frame of experimental conditons. See Section \code{details} for 
instructions how to create this object.}

\item{data}{can be either a list of data frames featuring a data frame with data
for each experiment or can be NULL when filepaths for the respective experiments are indicated in 
the configTable}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the suffix \code{fcStr} 
will be regarded as containing fold change values.}

\item{addCol}{additional column names that specify columns in the input data that are 
to be attached to the data frame throughout the analysis}

\item{intensityStr}{character string indicating which columns contain the sumionarea values.}

\item{qualColName}{character string indicating which column can be used for 
additional quality criteria when deciding between different non-unique 
protein identifiers.}
}
\value{
A list of data frames in which the experimental data are stored row-wise for each
  protein.
}
\description{
Creates a 2D-TPP data frame-list featuring a dataframe for each 
  temperature analyzed in the experiment.
}
\details{
Invokes the following steps: \enumerate{ \item  either reads in temperature specific 
  experimental data and creates data frame or extracts temperature specific data from pre-
  existing data frame list}
}

