% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dEvalConfigTable.R
\name{tpp2dEvalConfigTable}
\alias{tpp2dEvalConfigTable}
\title{Evaluation of 2D-TPP Configuration File}
\usage{
tpp2dEvalConfigTable(configTable)
}
\arguments{
\item{configTable}{data frame or character object with the path to a file, 
that specifies important details of the 2D-TPP experiment. See Section 
\code{details} for instructions how to create this object}
}
\value{
A configtable that works with the 2D-TPP workflow
}
\description{
Evaluates whether the configuration file is handed over as data frame or as file path 
  and loads the file path if necessary
}
\examples{
data("panobinostat_2DTPP_smallExample")
configTable <- tpp2dEvalConfigTable(panobinostat_2DTPP_config)

configTable <- tpp2dEvalConfigTable(system.file("example_data/2D_example_data/hdac2D_config.txt", 
                                                package = "TPP"))

}

