% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dExport.R
\name{tpp2dExport}
\alias{tpp2dExport}
\title{Produce Excel table of 2D-TPP experiment.}
\usage{
tpp2dExport(configTable = NULL, tab = NULL, resultPath = NULL,
  idVar = "gene_name", fcStr = "rel_fc_", intensityStr = "signal_sum_",
  addCol = NULL, normalizedData = FALSE, trRef = NULL)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{tab}{Table with results of the 2D-TPP analysis.}

\item{resultPath}{path for storing results table}

\item{idVar}{character string indicating how the column that contains the unique protein 
identifiers is called}

\item{fcStr}{character string indicating how columns that contain the actual 
fold change values are called}

\item{intensityStr}{character string indicating how columns that contain the raw masspec signal 
(e.g. sumionareas) values are called}

\item{addCol}{additional names of columns which are to be attached to the result table}

\item{normalizedData}{boolean variable indicating whether the data has been normalized}

\item{trRef}{character string containing a valid system path to a TPP-TR reference RData
file}
}
\value{
Creates excel file of the TPP-CCR analysis of the 2D-TPP data.
}
\description{
Produce Excel table of 2D-TPP experiment analysis results.
}
\examples{
data("panob2D_isobQuant_example")
cfg <- panobinostat_2DTPP_config
datRaw <- panobinostat_2DTPP_data
data2d <- tpp2dImport(cfg, datRaw, fcStr = NULL)
fcData2d <- tpp2dComputeFoldChanges(cfg, data2d)
normData2d <- tpp2dNormalize(cfg, fcData2d)
config_ccr <- tpp2dCreateCCRConfigFile(cfg)
ccr2dResults <- tpp2dCurveFit(config_ccr, normData2d)
tpp2dExport(cfg, ccr2dResults, resultPath = getwd())

}

