% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dExportPlots.R
\name{tpp2dExportPlots}
\alias{tpp2dExportPlots}
\title{Export plots for 2D-TPP experiment.}
\usage{
tpp2dExportPlots(plotList, resultPath, type = "none")
}
\arguments{
\item{plotList}{list of ggplots returned from one of the plotting functions}

\item{resultPath}{path for storing results}

\item{type}{character string specifying which type of plot is to be exported}
}
\value{
None
}
\description{
Exports plots into plots/ directory in the resultPath
}
\details{
Creates pdf files of the afore created plots by \code{plotSplines},
\code{tpp2dPlotCCRAllCurves}, \code{tpp2dPlotCCRGoodCurves}, 
\code{tpp2dPlotCCRSingleCurves}
}

