% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dMerge2dRef.R
\name{tpp2dMerge2dRef}
\alias{tpp2dMerge2dRef}
\title{Merge 2D-TPP result data with TPP-TR reference data}
\usage{
tpp2dMerge2dRef(data.table = NULL, trRef = NULL, idVar = "representative")
}
\arguments{
\item{data.table}{dataframe containing the 2D-TPP results}

\item{trRef}{character string of a valid system path to a TPP-TR reference 
RData object}

\item{idVar}{character string matching the column containing the unique protein 
identifiers}
}
\value{
A data frame with results merged from 2D-TPP and TPP-TR reference
}
\description{
Merges 2D-TPP result data with TPP-TR reference data to generate
a big table including both results
}

