% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dPlotCCRAllCurves.R
\name{tpp2dPlotCCRAllCurves}
\alias{tpp2dPlotCCRAllCurves}
\title{Plot all 2D-TPP CCR curves}
\usage{
tpp2dPlotCCRAllCurves(configTable = NULL, data = NULL,
  idVar = "gene_name", fcStr = "rel_fc_", verbose = FALSE)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{data}{data frame returned by the \code{tpp2dCurveFit} function}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the suffix \code{fcStr} 
will be regarded as containing fold change values.}

\item{verbose}{boolean variable stating whether a print description of problems/success for 
plotting of each protein should be printed}
}
\value{
A list of all dose-response plots that could be fitted
}
\description{
Generates a list of plots for all proteins with all curves for the different 
temperatures of for a certain 2D-TPP experiment.
}
\examples{
data("panob2D_isobQuant_example")
cfg <- panobinostat_2DTPP_config
datRaw <- panobinostat_2DTPP_data
data2d <- tpp2dImport(cfg, datRaw, fcStr = NULL)
fcData2d <- tpp2dComputeFoldChanges(cfg, data2d)
normData2d <- tpp2dNormalize(cfg, fcData2d)
config_ccr <- tpp2dCreateCCRConfigFile(cfg)
ccr2dResults <- tpp2dCurveFit(config_ccr, normData2d)
allCurves <- tpp2dPlotCCRAllCurves(cfg, ccr2dResults,
                                     fcStr = "norm_rel_fc_")
allCurves[["HDAC2"]]


}

