% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dPlotQCpEC50.R
\name{tpp2dPlotQCpEC50}
\alias{tpp2dPlotQCpEC50}
\title{Plot quality control pEC50 plots}
\usage{
tpp2dPlotQCpEC50(resultTable = NULL, resultPath = NULL, trRef = NULL,
  idVar = "representative")
}
\arguments{
\item{resultTable}{data.frame containing the results of a CCR analysis of 2D-TPP data}

\item{resultPath}{character string containing a valid system path to which the the qc
plots will be written}

\item{trRef}{character string with a link to a TPP-TR reference object RData file}

\item{idVar}{character string indicating how the column that contains the unique protein 
identifiers is called}
}
\value{
A folder with plots for each identified protein that compare melting points in
 the reference data set with the 2D-TPP data set
}
\description{
Plots quality control plots which indicate at which temperatures the pEC50 
 values of the treatment curves lie in comparison to those of the reference data
}

