% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dRemoveNAs.R
\name{tpp2dRemoveNAs}
\alias{tpp2dRemoveNAs}
\title{Remove NAs in 2D-TPP data}
\usage{
tpp2dRemoveNAs(data.list)
}
\arguments{
\item{data.list}{list of data frames that contain the data for the 2D-TPP experiment after 
computation of fold changes}
}
\value{
A dataframe without any NA values
}
\description{
Removes NAs in the fold change columns of the 2D-TPP data
}

