% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dRemoveZeroSias.R
\name{tpp2dRemoveZeroSias}
\alias{tpp2dRemoveZeroSias}
\title{Remove rows with zero sumionarea values}
\usage{
tpp2dRemoveZeroSias(configTable, data.list,
  intensityStr = "sumionarea_protein_")
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment.}

\item{data.list}{list of data frames of corresponding experiment data}

\item{intensityStr}{character string indicating which columns contain the sumionarea 
values. Those column names containing the suffix \code{intensityStr} 
will be regarded as containing sumionare values.}
}
\value{
A list of data frames with NAs instead of zeros.
}
\description{
Removes zero sumionare values in a specified data.list so that no errors are
  generated in the following fold change computation step. A corresponding data.list  
  with NAs instead of zeros is returned.
}

