% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dReplaceColNames.R
\name{tpp2dReplaceColNames}
\alias{tpp2dReplaceColNames}
\title{Replace column names for 2D-TPP data}
\usage{
tpp2dReplaceColNames(configTable, data.list, intensityStr, fcStr)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{data.list}{list of data frames that contain the data for the 2D-TPP experiment}

\item{intensityStr}{character string indicating which columns contain the actual 
sumionarea values. Those column names containing the suffix \code{intensityStr} 
will be regarded as containing sumionarea values.}

\item{fcStr}{character string indicating which columns contain the fold changes}
}
\value{
A list of dataframe with colnames which match concentrations instead of isobaric 
 labels
}
\description{
Replaces column names for 2D-TPP data so that the TPP-CCR main function can 
  deal with the data
}

