% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dRunTPPCCR.R
\name{tpp2dRunTPPCCR}
\alias{tpp2dRunTPPCCR}
\title{Run TPP-CCR analysis for 2D-TPP experiment}
\usage{
tpp2dRunTPPCCR(configFile, data, nCores = 1, naStrs = c("NA", "n/d", "NaN",
  "<NA>"), fcStr = "norm_rel_fc_protein_", idVar = NULL,
  nonZeroCols = "qupm", r2Cutoff = 0.8, fcCutoff = 1.5,
  slopeBounds = c(1, 50), fcTolerance = 0.1)
}
\arguments{
\item{configFile}{list of dataframes, that specifies important details of the 2D-TPP 
experiment for each temperature.}

\item{data}{data.frame, that contains the data of the 2D-TPP 
experiment for each temperature.}

\item{nCores}{numeric value stating how many cores are to be used for computation}

\item{naStrs}{character vector indicating missing values in the data table. 
When reading data from file, this value will be passed on to the argument}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the suffix \code{fcStr} 
will be regarded as containing fold change values.}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}

\item{nonZeroCols}{character string indicating a column that will be used for
filtering out zero values.}

\item{r2Cutoff}{Quality criterion on dose response curve fit.}

\item{fcCutoff}{Cutoff for highest compound concentration fold change.}

\item{slopeBounds}{Bounds on the slope parameter for dose response curve 
fitting.}

\item{fcTolerance}{tolerance for the fcCutoff parameter. See details.}
}
\value{
A data frames in which the fit results are stored row-wise for each
  protein.
}
\description{
Performs analysis of a TPP-CCR experiment by invoking the routine 
  for TPP-CCR curve fitting for each temperature of the sample.
}
\examples{
  load(system.file("example_data/2D_example_data/referenceCCRConfig.RData", package="TPP"))
  load(system.file("example_data/2D_example_data/exampleRunCCRInput.RData", package="TPP"))
  CCRresults <- tpp2dRunTPPCCR(configFile=exampleCCRConfig, data=exampleRunCCRInput, 
                               idVar="unique_ID")
  
}

