% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dSplineFitAndTest.R
\name{tpp2dSplineFitAndTest}
\alias{tpp2dSplineFitAndTest}
\title{Fit splines and perform f-Test}
\usage{
tpp2dSplineFitAndTest(data_2D, trRefDataPath, idVar = "representative",
  fcStr = "norm_rel_fc_protein_", refFcStr = "norm_rel_fc_protein_",
  resultPath = NULL, ggplotTheme = tppDefaultTheme(), doPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{data_2D}{result data.frame from a 2D-TPP CCR analysis}

\item{trRefDataPath}{character string with a link to a TPP-TR reference object RData file}

\item{idVar}{character string indicating name of the columns containing the unique protein 
identifiers}

\item{fcStr}{character string indicating how columns that will contain the actual 
fold change values will be called. The suffix \code{fcStr} will be pasted in front of
the names of the experiments.}

\item{refFcStr}{same as argument fcStr, but for the reference data.}

\item{resultPath}{location where to store dose-response curve plots and 
results table.}

\item{ggplotTheme}{ggplot theme for protein-wise plots.}

\item{doPlot}{boolan value indicating whether protein-wise plots should be 
produced Deactivating plotting decreases runtime.}

\item{verbose}{print description of problems for each protein for which splines fits could 
not be performed}
}
\value{
None
}
\description{
Fit splines through TR reference dataset and extrapolates relative 2D-TPP datapoints, 
 then compares spline fits of different treatments with non-treatment with an f-test
}

