% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dTRReferenceObject.R
\name{tpp2dTRReferenceObject}
\alias{tpp2dTRReferenceObject}
\title{TPP-TR reference object}
\usage{
tpp2dTRReferenceObject(
  tppRefData = NULL,
  tppRefPath = NULL,
  fcStr = "norm_rel_fc_",
  qualColName = "qupm"
)
}
\arguments{
\item{tppRefData}{TPP-TR reference object that can be directly passed to the function}

\item{tppRefPath}{character string containing a system path to a RData file containing an
TPP-TR reference object}

\item{fcStr}{character string indicating which columns contain the fold changes}

\item{qualColName}{character string indicating which column contain protein 
identification quality measures}
}
\value{
A TPP-TR reference object
}
\description{
Definition of a TPP-TR reference object
}
\examples{
trRef <- system.file("example_data/2D_example_data/referenceNormData.RData", package="TPP")
tpp2dTRReferenceObject(tppRefPath=trRef)

}
