% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tppccrImport.R
\name{tppccrImport}
\alias{tppccrImport}
\title{Import TPP-CCR dataset for analysis by the 
  \code{\link[TPP]{TPP}} package.}
\usage{
tppccrImport(
  configTable,
  data = NULL,
  idVar = "gene_name",
  fcStr = "rel_fc_",
  naStrs = c("NA", "n/d", "NaN", "<NA>"),
  qualColName = "qupm",
  nonZeroCols = "qssm"
)
}
\arguments{
\item{configTable}{either a dataframe or the path to a spreadsheet. In both cases
it specifies necessary information of the TPP-CCR experiment.}

\item{data}{dataframe containing fold change measurements and 
additional annotation columns to be imported. Can be used instead of 
specifying the file path in \code{configTable}.}

\item{idVar}{character string indicating which data column provides the unique 
identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the suffix 
\code{fcStr} will be regarded as containing fold change values.}

\item{naStrs}{character vector indicating missing values in the data table. 
When reading data from file, this value will be passed on to the argument 
\code{na.strings} in function \code{read.delim}.}

\item{qualColName}{character string indicating which column can be used for 
additional quality criteria when deciding between different non-unique 
protein identifiers.}

\item{nonZeroCols}{character string indicating a column that will be used for
filtering out zero values.}
}
\value{
ExpressionSet object storing the measured fold changes, as well as
row and column metadata. In each ExpressionSet \code{S}, the fold changes can 
be accessed by \code{Biobase::exprs(S)}. Protein expNames can be accessed by 
\code{featureNames(S)}. Isobaric labels and the corresponding concentrations are 
  returned by \code{S$label} and \code{S$concentration}.
}
\description{
\code{tppccrImport} imports a table of protein fold changes and 
stores them in an ExpressionSet for use in the \code{\link[TPP]{TPP}} package.
}
\details{
The imported dataset has to contain measurements obtained by a
  TPP-CCR experiment. Fold changes need to be pre-computed using the lowest 
  concentration as reference.
  
  The dataset can be specified by filename in the \code{configTable} 
  argument, or given directly in the \code{data} argument
  
  The default settings are adjusted to analyze data of the python package 
  \code{isobarQuant}. You can also customize them for your own dataset.
  
  The \code{configTable} argument is a dataframe, or the path to a 
  spreadsheet (tab-delimited text-file without quoted strings, or xlsx format). 
  Information about each experiment is stored row-wise. 
  It contains the following columns: 
  \describe{
 \item{\code{Path}}{location of the datafile. Alternatively, data can be directly handed
  over by the \code{data} argument.}
 \item{\code{Experiment}}{unique experiment name.}
 \item{Label columns}{each isobaric label names a column that contains the 
  concentration administered for the label in the individual experiments.
}
}
  
  During data import, proteins with NAs in the data column specified by \code{idVar} receive 
  unique generic IDs so that they can be processed by the package.
}
\examples{
data(hdacCCR_smallExample)
tppccrData <- tppccrImport(configTable=hdacCCR_config, 
data = hdacCCR_data)

}
\seealso{
\code{\link{tpptrImport}}, \code{\link{tppccrCurveFit}}
}
