% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tppExport.R
\name{tppExport}
\alias{tppExport}
\title{Produce Excel table of TPP-TR or TPP-CCR experiment.}
\usage{
tppExport(tab, file)
}
\arguments{
\item{tab}{Table with results of the TPP analysis.}

\item{file}{path for storing results table}
}
\value{
No value returned.
}
\description{
Produce Excel table of TPP-TR or TPP-CCR experiment.
}
\examples{
data(hdacTR_resultsTable_smallExample)
tppExport(resultTable, "tpptr_example_results.xlsx")
}

