% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPPTR_reference_results_HepG2.R
\name{tppRefData}
\alias{tppRefData}
\title{Example of a reference dataset for 2D-TPP experiments.}
\description{
Reference dataset obtained by TPP-TR experiments without drug
treatment on HepG2 cell lines.
}
\details{
\code{tppRefData} is a list of data frames that contains
  TPP-TR measurements for a large number of proteins in wide format.
  The experiments were performed in two replicates. It can be used as a 
  reference for normalization of 2D-TPP data. See the vignette for the 2D
  workflow for details.
}

