% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPPTR_reference_results_HepG2.R
\docType{data}
\name{TPPTR_reference_results_HepG2}
\alias{TPPTR_reference_results_HepG2}
\title{Example of a reference dataset for 2D-TPP experiments.}
\description{
Reference dataset obtained by TPP-TR experiments without drug
treatment on HepG2 cell lines.
}
\details{
Contains the data object \code{\link{tppRefData}}.
}

