% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tpptrDefaultNormReqs.R
\name{tpptrDefaultNormReqs}
\alias{tpptrDefaultNormReqs}
\title{Default filter criteria for fold change normalization}
\usage{
tpptrDefaultNormReqs()
}
\value{
List with two entries: 'fcRequirements' describes filtering
requirements on fold change columns, 'otherRequirements' contains criteria on
additional metadata columns.
}
\description{
Filter criteria as described in the publication.
}
\examples{
data(hdacTR_smallExample)
tpptrData <- tpptrImport(configTable=hdacTR_config, data=hdacTR_data)
tpptrNorm <- tpptrNormalize(data=tpptrData, normReqs=tpptrDefaultNormReqs())
}

