% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tpptrQCPlotsMinSlopes_vs_MPdiffs.R
\name{tpptrQCPlotsMinSlopes_vs_MPdiffs}
\alias{tpptrQCPlotsMinSlopes_vs_MPdiffs}
\title{Plot minSlope vs. melting point differences}
\usage{
tpptrQCPlotsMinSlopes_vs_MPdiffs(resultTable, expNames = NULL,
  ggplotTheme = tppDefaultTheme(), expRepl = NULL, expCond = NULL,
  path = NULL)
}
\arguments{
\item{resultTable}{Table of the TPP-TR results.}

\item{expNames}{Names of each experiment.}

\item{ggplotTheme}{ggplot theme to be used in the plots.}

\item{expRepl}{Replicate each experiment belongs to.}

\item{expCond}{Condition (Treatment or Vehicle) each experiment belongs to.}

\item{path}{Location where to store resulting plot.}
}
\value{
No value returned.
}
\description{
\code{tpptrQCPlotsMinSlopes_vs_MPdiffs} plots the minSlope vs.
  melting point difference for each protein and highlights proteins with
  significant shifts.
}
\examples{
data(hdacTR_resultsTable_smallExample)
tpptrQCPlotsMinSlopes_vs_MPdiffs(resultTable=resultTable,
expNames=c("Vehicle_1", "Vehicle_2", "Panobinostat_1", "Panobinostat_2"),
expRepl=c(1,2,1,2),
expCond=c("Vehicle", "Vehicle", "Treatment", "Treatment"))
}

