% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tpptrResultTable.R
\name{tpptrResultTable}
\alias{tpptrResultTable}
\title{Statistical analysis of melting curve parameters.}
\usage{
tpptrResultTable(data, binWidth = 300)
}
\arguments{
\item{data}{list of ExpressionSets containing fold changes and metadata. It's
featureData contains the fitted melting curve parameters}

\item{binWidth}{bin width used for p-value computation}
}
\value{
A data frame in which the fit results are stored row-wise for each
  protein.
}
\description{
Summarizes the output of a TPP-TR experiment and performs
  statistical comparisons of conditions, if appropriate.
}
\details{
If a TPP-TR experiment was performed with the conditions "Vehicle"
  and "Treatment", the melting points between these conditions will be
  compared statistically, producing p-values for each protein and replicate.
}
\examples{
data(hdacTR_fittedData_smallExample)
resultTable <- tpptrResultTable(trDataFitted)
subset(resultTable, fulfills_all_4_requirements)$Protein_ID
}
\references{
Cox, J., & Mann, M. (2008). MaxQuant enables high peptide
  identification rates, individualized ppb-range mass accuracies and
  proteome-wide protein quantification. Nature biotechnology, 26(12),
  1367-1372.
}

