% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppVenn.R
\name{tppVenn}
\alias{tppVenn}
\title{Venn diagrams of detected proteins per experiment.}
\usage{
tppVenn(data)
}
\arguments{
\item{data}{list of ExpressionSets that contain the imported data per
experiment (return value of function \code{\link{tpptrImport}} or
\code{\link{tppccrImport}}.}
}
\value{
Venn diagram plot. Can be plotted by \code{\link{grid.draw}}.
}
\description{
\code{tppVenn} illustrates the overlaps between different
  TPP-TR/CCR experiments by a venn diagram.
}
\examples{
data(hdacTR_smallExample)
trImported <- tpptrImport(configTable=hdacTR_config, data=hdacTR_data)
vennPlot <- tppVenn(data=trImported)
grid.draw(vennPlot)
}

