% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdacTR_fittedData_smallExample.R
\name{trDataFitted}
\alias{trDataFitted}
\title{Example of a TPP-TR dataset with fitted melting curve parameters.}
\description{
Example of a TPP-TR dataset with fitted melting curve parameters.
}
\details{


\code{trDataFitted} is a list of ExpressionSets that contain the
 measurements of a TPP-TR experiment (accessible by the \code{exprs}
 command), as well as fitted melting curve parameters in the
 \code{featureData}. It can be used as input for the function
 \code{\link{tpptrResultTable}}.
}

