% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze2DTPP.R
\name{analyze2DTPP}
\alias{analyze2DTPP}
\title{Analyze 2D-TPP experiment}
\usage{
analyze2DTPP(configFile = NULL, data = NULL, resultPath = NULL,
  idVar = "representative", fcStr = "rel_fc_protein_",
  intensityStr = "sumionarea_protein_", naStrs = c("NA", "n/d", "NaN",
  "<NA>"), methods = c("doseRespone", "splineFit"), qualColName = "qupm",
  compFc = TRUE, normalize = TRUE, addCol = NULL, nCores = "max",
  nonZeroCols = "qupm", fcTolerance = 0.1, r2Cutoff = 0.8,
  fcCutoff = 1.5, slopeBounds = c(1, 50), fractAbund = FALSE,
  xlsxExport = FALSE, plotAll = FALSE, plotAllR2 = FALSE,
  plotSingle = FALSE, trRef = NULL, refFcStr = "norm_rel_fc_protein_",
  addInfo = FALSE, createReport = "html_document")
}
\arguments{
\item{configFile}{dataframe, or character object with the path to a file, 
that specifies important details of the 2D-TPP experiment. See Section 
\code{details} for instructions how to create this object.}

\item{data}{single dataframe, containing fold change measurements and 
additional annotation columns to be imported. Can be used instead of 
specifying the file path in the \code{configTable} argument.}

\item{resultPath}{location where to store dose-response curve plots and 
results table.}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the prefix \code{fcStr} 
will be regarded as containing fold change values.}

\item{intensityStr}{character string indicating which columns contain the actual 
sumionarea values. Those column names containing the prefix \code{intensityStr} 
will be regarded as containing sumionarea values.}

\item{naStrs}{character vector indicating missing values in the data table. 
When reading data from file, this value will be passed on to the argument 
\code{na.strings} in function \code{read.delim}.}

\item{methods}{vector of character strings that indicate which methods should be used 
for the analysis (default: c("doseRespone"), alternative: c("splineFit") or 
c("doseRespone", "splineFit"))}

\item{qualColName}{character string indicating which column can be used for 
additional quality criteria when deciding between different non-unique 
protein identifiers.}

\item{compFc}{boolean flag which indicates whether to perform fold change computation 
regarding reference column from sumionareas (default: TRUE)}

\item{normalize}{perform median normalization (default: TRUE).}

\item{addCol}{vector of chracter strings indicating which additional columns to include 
from the input data}

\item{nCores}{either a numerical value given the desired number of CPUs, or 
'max' to automatically assign the maximum possible number (default).}

\item{nonZeroCols}{character string indicating a column that will be used for
filtering out zero values.}

\item{fcTolerance}{tolerance for the fcCutoff parameter. See details.}

\item{r2Cutoff}{Quality criterion on dose response curve fit.}

\item{fcCutoff}{Cutoff for highest compound concentration fold change.}

\item{slopeBounds}{Bounds on the slope parameter for dose response curve 
fitting.}

\item{fractAbund}{boolean variable, if set to TRUE additional information concerning
sumionarea fractional abundance and dmso1 vs. dmso2 of adjacent temperatures is 
added to the output table}

\item{xlsxExport}{produce results table in xlsx format and store at the 
location specified by the \code{resultPath} argument.}

\item{plotAll}{boolan value indicating whether all dose response curves should
be generated. Deactivating plotting decreases runtime.}

\item{plotAllR2}{boolan value indicating whether all dose response curves which
fullfill the demanded criterias (Rsquared, maximum plateau) should be generated. 
Deactivating plotting decreases runtime.}

\item{plotSingle}{boolan value indicating whether all dose response curves which
fullfill the demanded criterias (Rsquared, maximum plateau) should be generated. 
Deactivating plotting decreases runtime.}

\item{trRef}{character string containing a valid system path to a previously generated TPP-TR
reference object}

\item{refFcStr}{character string indicating which columns in the reference data set contain 
the fold change values}

\item{addInfo}{boolean variable, if set to TRUE additional information on counts of
stabilization and destabiliazation of each protein is added to the output table}

\item{createReport}{character string indicating whether a markdown report should be created
and which format it have (default: "html_document", alternative: "pdf_document" or "none")}
}
\value{
A data frame in which the fit results are stored row-wise for each
  protein at the different temperatures.
}
\description{
Performs analysis of a 2D-TPP experiment by invoking routines 
  for data import, data processing, fold change computation, median normalization, 
  TPP-CCR curve fitting, plotting and production of the result table.
}
\details{
Invokes the following steps: \enumerate{ \item Import data using the
  \code{\link{tpp2dImportData}} function. \item Remove zero sumionarea values. 
  \item Compute fold changes from raw data (sumionarea) 
  \item Perform normalization by fold 
  change medians (optional) using the \code{\link{tpp2dDoMedianNorm}} function.
  To perform normalization, set argument \code{normalize=TRUE}.}
}
\examples{
data("panobinostat_2DTPP_smallExample")
config_tpp2d <- panobinostat_2DTPP_config
data_tpp2d <- panobinostat_2DTPP_data
tpp2dResults <- analyze2DTPP(configFile = config_tpp2d, 
                             data = data_tpp2d,
                             fcStr = NULL,
                             methods=c("doseResponse"),
                             createReport="none",
                             nCores=1)

}
\references{
Becher, I., Werner, T., Doce, C., Zaal, E. A., Berkers, C. R., T"ogel, I., 
Salzer, E., Bantscheff, M., Savitski, M. M. (2016) Comprehensive thermal and chemoproteomics 
profiling identifies phenylalanine hydroxylase as a potent off-target of the histone 
deacetylase inhibitor panobinostat. Nature Chemical Biology (accepted)
}

