% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\name{hdacCCR_data}
\alias{hdacCCR_data}
\title{TPP-CCR example dataset (replicates 1 and 2)}
\description{
Example subset of a Panobinostat TPP-CCR dataset (replicates 1 
and 2)
}
\details{
A list with two subsets of a dataset obtained by TPP-CCR experiments to 
  investigate drug effects for HDAC inhibitor Panobinostat. It contains 7 
  HDACs as well as a random selection of 493 further proteins.
  
  You can use this dataset to explore the \code{\link{TPP}} package 
  functionalities without invoking the whole time consuming analysis on the 
  big dataset.
  
  The original dataset is located in the folder 
  \code{'example_data/CCR_example_data'} in the package's installation 
  directory. You can find it on your system by the \code{R} command 
  \code{system.file('example_data', package = 'TPP')}.
}

