% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\docType{data}
\name{hdacCCR_smallExample}
\alias{hdacCCR_smallExample}
\title{Example subsets of a Panobinostat TPP-CCR dataset (replicates 1 and 2) and the
  corresponding configuration table to start the analysis.}
\description{
Example dataset obtained by TPP-CCR experiments for analysis by 
the TPP-package. It contains all necessary arguments to start the analysis
(config table and list of data frames).
}

