% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacCCR_smallExample.R
\docType{data}
\name{hdacCCR_smallExample}
\alias{hdacCCR_smallExample}
\title{Example subsets of a Panobinostat TPP-CCR dataset (replicates 1 and 2) and the
  corresponding configuration tables to start the analysis.}
\description{
Example subsets of a Panobinostat TPP-CCR dataset (replicates 1 and 2) and the
  corresponding configuration tables to start the analysis.
}

