% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\name{hdacTR_config}
\alias{hdacTR_config}
\title{The configuration table to analyze \link{hdacTR_data}.}
\description{
The configuration table to analyze \link{hdacTR_data}.
}
\details{


\code{hdacTR_config} is a data frame that specifies the experiment
 name, isobaric labels, and the administered drug concentrations at each
 label.
}

