% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\name{hdacTR_data}
\alias{hdacTR_data}
\title{TPP-TR example dataset.}
\description{
Example subset of a dataset obtained by TPP-TR experiments to 
  investigate possible targets for HDAC inhibitor Panobinostat.
}
\details{
\code{hdacTR_data} is a list of data frames that contain
  measurements for HDACs as well as a random selection of 500 further
  proteins.
  
  You can use this dataset to explore the \code{\link{TPP}} package 
  functionalities without invoking the whole time consuming analysis on the 
  whole dataset.
  
  The originial dataset is located in the folder 
  \code{'example_data/TR_example_data'} in the package's installation
  directory. You can find it on your system by the \code{R} command
  \code{system.file('example_data', package = 'TPP')}.
}

