% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdacTR_smallExample.R
\docType{data}
\name{hdacTR_smallExample}
\alias{hdacTR_smallExample}
\title{Example subset of a Panobinostat TPP-TR dataset and the
  corresponding configuration table to start the analysis.}
\description{
Example subset of a Panobinostat TPP-TR dataset and the
  corresponding configuration table to start the analysis.
}

