% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panob2D_isobQuant_example.R
\name{panobinostat_2DTPP_data}
\alias{panobinostat_2DTPP_data}
\title{2D-TPP-CCR example dataset}
\description{
Example subset of a Panobinostat 2D-TPP dataset
}
\details{
A list with two subsets of a dataset obtained by 2D-TPP experiments to 
  investigate drug effects for HDAC inhibitor Panobinostat. The experiment 
  was performed on living HepG2 cells (see Becher et al. (2016). Thermal 
  profiling reveals phenylalanine hydroxylase as an off-target of panobinostat. 
  Nature Chemical Biology, (September)) 
  It contains 7 HDACs as well as a random selection of 493 further proteins.
  
  You can use this dataset to explore the \code{\link{TPP}} package 
  functionalities without invoking the whole time consuming analysis on the 
  big dataset.
  
  The original dataset was not produced by isobarQuant but by a custom-made
  software instead. For illustrative purposes, the column names have been 
  adapted to the isobarQuant convention (id column = "gene_name", 
  signal columns starting with = "signal_sum_", 
  quantified spectrum matches = "qusm", quantified peptide matches = "qupm").
  The original unmodified dataset in plain text format is still located in the 
  folder  \code{'example_data/2D_example_data'} in the package's installation 
  directory. You can find it on your system by the \code{R} command 
  \code{system.file('example_data', package = 'TPP')}.
}

