% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panobinostat_2DTPP_smallExample.R
\name{panobinostat_2DTPP_data}
\alias{panobinostat_2DTPP_data}
\title{2D-TPP-CCR example dataset}
\description{
Example subset of a Panobinostat 2D-TPP dataset
}
\details{
A list with two subsets of a dataset obtained by 2D-TPP experiments to 
  investigate drug effects for HDAC inhibitor Panobinostat. The experiment 
  was performed on living HepG2 cells (see Becher et al. (2016). Thermal 
  profiling reveals phenylalanine hydroxylase as an off-target of panobinostat. 
  Nature Chemical Biology, (September)) 
  It contains 7 HDACs as well as a random selection of 493 further proteins.
  
  You can use this dataset to explore the \code{\link{TPP}} package 
  functionalities without invoking the whole time consuming analysis on the 
  big dataset.
}
\seealso{
\code{\link{panobinostat_2DTPP_config}, \link{panobinostat_2DTPP_smallExample}}
}
