% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panobinostat_2DTPP_smallExample.R
\docType{data}
\name{panobinostat_2DTPP_smallExample}
\alias{panobinostat_2DTPP_smallExample}
\title{Example subsets of a Panobinostat 2D-TPP dataset and the
  corresponding configuration table to start the analysis.}
\description{
Example dataset obtained by 2D-TPP experiments for analysis by 
the TPP-package. It contains all necessary arguments to start the analysis
(config table and list of data frames).
}
\seealso{
\code{\link{panobinostat_2DTPP_data}, \link{panobinostat_2DTPP_config}}
}

