% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tppDeprecated.R
\name{TPP-deprecated}
\alias{TPP-deprecated}
\alias{tpp2dCreateCCRConfigFile}
\alias{tpp2dEvalConfigTable}
\alias{tpp2dPlotCCRAllCurves}
\alias{tpp2dPlotCCRGoodCurves}
\alias{tpp2dPlotCCRSingleCurves}
\alias{tpp2dRemoveZeroSias}
\alias{tpp2dReplaceColNames}
\title{Deprecated functions in package \sQuote{TPP}}
\usage{
tpp2dPlotCCRGoodCurves(configTable = NULL, data = NULL,
  idVar = "gene_name", fcStr = "rel_fc_", verbose = FALSE)

tpp2dPlotCCRAllCurves(configTable = NULL, data = NULL,
  idVar = "gene_name", fcStr = "rel_fc_", verbose = FALSE)

tpp2dPlotCCRSingleCurves(configTable = NULL, data = NULL,
  idVar = "gene_name", fcStr = "rel_fc_", verbose = FALSE)

tpp2dEvalConfigTable(configTable)

tpp2dRemoveZeroSias(configTable, data.list, intensityStr = "signal_sum_")

tpp2dReplaceColNames(configTable, data.list, intensityStr, fcStr)

tpp2dCreateCCRConfigFile(configTable)
}
\arguments{
\item{configTable}{DEPRECATED}

\item{data}{DEPRECATED}

\item{idVar}{DEPRECATED}

\item{fcStr}{DEPRECATED}

\item{verbose}{DEPRECATED}

\item{data.list}{DEPRECATED}

\item{intensityStr}{DEPRECATED}
}
\value{
No value returned
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{TPP} only, and will be defunct at the next release.
}
\details{
The following function is deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{
 \item{tpp2dPlotCCRGoodCurves: \code{\link{tpp2dCreateDRplots}}}
 \item{tpp2dPlotCCRSingleCurves: \code{\link{tpp2dCreateDRplots}}}
 \item{tpp2dPlotCCRAllCurves: \code{\link{tpp2dCreateDRplots}}}
 }
}

