% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dCalcFractAbundance.R
\name{tpp2dCalcFractAbundance}
\alias{tpp2dCalcFractAbundance}
\title{Calculate fractional abundance and DMSO ratio of successive sumionareas  
  (usage of function is only reasonable when at least two temperatures are multiplexed!)}
\usage{
tpp2dCalcFractAbundance(configTable = NULL, data = NULL,
  intensityStr = NULL, idVar = NULL)
}
\arguments{
\item{configTable}{data frame that specifies important details of the TPP-CCR experiment.}

\item{data}{data frame of TPP-CCR results (e.g. obtained by \code{run2DTPPCCR}).}

\item{intensityStr}{character string indicating which columns contain the sumionarea values.}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}
}
\value{
Data frame that was handed over with additional columns of fractional abundance
  and DMSO1 vs DMSO2 ratio
}
\description{
Calculates fractional abundance and DMSO ratio of successive sumionareas and 
  creates respective columns which are added two the data frame which is handed over
}
\examples{
  data("panobinostat_2DTPP_smallExample")
  load(system.file("example_data/2D_example_data/fractAbundInData.RData", package="TPP"))
  fracAbund <- tpp2dCalcFractAbundance(configTable = panobinostat_2DTPP_config, 
                data = fractAbundInData,
                intensityStr = "sumionarea_protein_", 
                idVar = "representative")
 
}

