% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dExport.R
\name{tpp2dExport}
\alias{tpp2dExport}
\title{Produce Excel table of 2D-TPP experiment.}
\usage{
tpp2dExport(configTable = NULL, tab = NULL, outPath = NULL,
  idVar = NULL, fcStr = NULL, intensityStr = NULL, addCol = NULL,
  normalizedData = FALSE, trRef = NULL)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{tab}{Table with results of the 2D-TPP analysis.}

\item{outPath}{path for storing results table}

\item{idVar}{character string indicating how the column that contains the unique protein 
identifiers is called}

\item{fcStr}{character string indicating how columns that contain the actual 
fold change values are called}

\item{intensityStr}{character string indicating how columns that contain the raw masspec signal 
(e.g. sumionareas) values are called}

\item{addCol}{additional names of columns which are to be attached to the result table}

\item{normalizedData}{boolean variabel indicating whether the data has been normalized}

\item{trRef}{character string containing a valid system path to a TPP-TR reference RData
file}
}
\value{
Creates excel file of the TPP-CCR analysis of the 2D-TPP data.
}
\description{
Produce Excel table of 2D-TPP experiment analysis results.
}
\examples{
data("panobinostat_2DTPP_smallExample")
load(system.file("example_data/2D_example_data/shortData2d.RData", package="TPP"))
# tpp2dExport(configTable = panobinostat_2DTPP_config, tab=shortData2d, 
#             outPath=getwd(), 
#             idVar="representative", fcStr="norm_rel_fc_protein_", 
#             intensityStr="sumionarea_protein_", addCol=NULL)

}

