% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dNormalize.R
\name{tpp2dNormalize}
\alias{tpp2dNormalize}
\title{Median normalization of protein fold changes of 2D-TPP data}
\usage{
tpp2dNormalize(configTable, data, fcStr = "rel_fc_protein_")
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{data}{data frame that contains the data for the 2D-TPP experiment}

\item{fcStr}{character string indicating how columns that will contain the actual 
fold change values will be called. The suffix \code{fcStr} will be pasted in front of
the names of the experiments.}
}
\value{
A dataframe identical to the input dataframe except that the columns containing the
  fold change values have been normalized by their median.
}
\description{
Normalizes fold changes retrieved from 2D-TPP experiment by dividing by the median fold
  change
}
\examples{
  load(system.file("example_data/2D_example_data/referenceFcData.RData", package="TPP"))
  data("panobinostat_2DTPP_smallExample")
  NormData2d <- tpp2dNormalize(configTable = panobinostat_2DTPP_config, 
                                    data=headData2dFc,
                                    fcStr="rel_fc_protein_")
  
}

