% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dPlotCCRAllCurves.R
\name{tpp2dPlotCCRAllCurves}
\alias{tpp2dPlotCCRAllCurves}
\title{Plot all 2D-TPP CCR curves}
\usage{
tpp2dPlotCCRAllCurves(configTable = NULL, data = NULL, idVar = NULL,
  fcStr = "rel_fc_protein_", verbose = FALSE)
}
\arguments{
\item{configTable}{data frame that specifies important details of the 2D-TPP experiment}

\item{data}{data frame returned by the \code{tpp2dRunTPPCCR} function}

\item{idVar}{character string indicating which data column provides the 
unique identifiers for each protein.}

\item{fcStr}{character string indicating which columns contain the actual 
fold change values. Those column names containing the suffix \code{fcStr} 
will be regarded as containing fold change values.}

\item{verbose}{boolean variable stating whether a print description of problems/success for 
plotting of each protein should be printed}
}
\value{
A list of all dose-reponse plots that could be fitted
}
\description{
Generates a list of plots for all proteins with all curves for the different 
temperatures of for a certain 2D-TPP experiment.
}
\examples{
data("panobinostat_2DTPP_smallExample")
cfg <- panobinostat_2DTPP_config
datRaw <- panobinostat_2DTPP_data
data2d <- tpp2dImportData(cfg, datRaw, fcStr = NULL)
fcData2d <- tpp2dComputeFoldChanges(cfg, data2d, intensityStr="sumionarea_protein_")
normData2d <- tpp2dDoMedianNorm(cfg, fcData2d)
config_ccr <- tpp2dCreateCCRConfigFile(cfg)
ccr2dResults <- tpp2dRunTPPCCR(config_ccr, normData2d, idVar = "unique_ID")
allCurves <- tpp2dPlotCCRAllCurves(cfg, ccr2dResults, 
                                     idVar = "representative",
                                     fcStr = "norm_rel_fc_protein_")
allCurves[["IPI00289601.10"]]


}

