% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppccrCurveFit.R
\name{tppccrCurveFit}
\alias{tppccrCurveFit}
\title{Fit and analyse dose response curves}
\usage{
tppccrCurveFit(data, resultPath = NULL, ggplotTheme = tppDefaultTheme(),
  doPlot = TRUE, fcCutoff = 1.5, r2Cutoff = 0.8, slopeBounds = c(1, 50))
}
\arguments{
\item{data}{ExpressionSet containing protein fold changes for curve fitting.}

\item{resultPath}{location where to store dose-response curve plots and
result table.}

\item{ggplotTheme}{ggplot theme for dose response curve plots.}

\item{doPlot}{boolan value indicating whether dose response curves should be
plotted. Deactivating plotting decreases runtime.}

\item{fcCutoff}{cutoff for highest compound concentration fold change.}

\item{r2Cutoff}{quality criterion on dose response curve fit.}

\item{slopeBounds}{bounds on the slope parameter for dose response curve
fitting.}
}
\value{
A data frame in which the fit results are stored row-wise for each
  protein.
}
\description{
\code{tppccrCurveFit} fits logistic dose response curves to fold
  change measurements of a TPP-CCR experiment and returns quality information
  about the estimated parameters.
}
\details{
\code{data} is an ExpressionSet object created by
\code{\link{tppccrImport}}. It contains the isobaric labels and administered
drug concentrations in the \code{phenoData} and user-defined protein
properties in the \code{featureData}. Protein IDs are stored in the
\code{featureNames}.

The dose response curve plots will be stored in a subfolder with name
\code{DoseResponse_Curves} at the location specified by \code{resultPath}.
}
\examples{
data(hdacCCR_smallExample)
tppccrData <- tppccrImport(configTable=hdacCCR_config_repl1,
                           data=hdacCCR_data_repl1)
tppccrNorm <- tppccrNormalize(data=tppccrData)
tppccrTransformed <- tppccrTransform(data=tppccrNorm)
tppccrFitResults <- tppccrCurveFit(data=tppccrTransformed)
subset(tppccrFitResults, passed_filter)
}
\seealso{
\code{\link{tppccrImport}}, \code{\link{tppDefaultTheme}}
}

