% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tppccrNormalize.R
\name{tppccrNormalize}
\alias{tppccrNormalize}
\title{Normalize data from TPP-CCR experiments}
\usage{
tppccrNormalize(data)
}
\arguments{
\item{data}{expressionSet with measurements to be normalized}
}
\value{
ExpressionSet object storing the normalized fold changes, as well as
  row and column metadata. In each ExpressionSet \code{S}, the fold changes
  can be accessed by \code{exprs(S)}. Protein expNames can be accessed by
  \code{featureNames(S)}. TMT labels and the corresponding temperatures are
  returned by \code{S$labels} and \code{S$temperatures}.
}
\description{
Normalize each fold change column by its median.
}
\examples{
data(hdacCCR_smallExample)
tppccrData <- tppccrImport(configTable=hdacCCR_config_repl1, data = hdacCCR_data_repl1)
tppccrNorm <- tppccrNormalize(data=tppccrData)
head(exprs(tppccrNorm))
}

