% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tppccrTransform.R
\name{tppccrTransform}
\alias{tppccrTransform}
\title{Transform fold changes of TPP-CCR experiment}
\usage{
tppccrTransform(data, fcCutoff = 1.5)
}
\arguments{
\item{data}{expressionSet object containing the data to be transformed.}

\item{fcCutoff}{Cutoff for highest compound concentration fold change.}
}
\value{
ExpressionSet object storing the transformed fold changes, as well as
  row and column metadata. In each ExpressionSet \code{S}, the fold changes
  can be accessed by \code{exprs(S)}. Protein expNames can be accessed by
  \code{featureNames(S)}. TMT labels and the corresponding temperatures are
  returned by \code{S$labels} and \code{S$temperatures}.
}
\description{
Transform fold changes of TPP-CCR experiment to prepare them for
  dose response curve fitting.
}
\examples{
data(hdacCCR_smallExample)
tppccrData <- tppccrImport(configTable=hdacCCR_config_repl1, data = hdacCCR_data_repl1)
tppccrNorm <- tppccrNormalize(data=tppccrData)
tppccrTransformed <- tppccrTransform(data=tppccrNorm)
determinedEffectTypes <- featureData(tppccrTransformed)$CompoundEffect
transformedData <- data.frame(exprs(tppccrTransformed),
                              Type=determinedEffectTypes)
}

