% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tppExport.R
\name{tppExport}
\alias{tppExport}
\title{Produce Excel table of TPP-TR or TPP-CCR experiment.}
\usage{
tppExport(tab, file, expNames = NULL, expColors = NULL)
}
\arguments{
\item{tab}{Table with results of the TPP analysis.}

\item{file}{path for storing results table}

\item{expNames}{character vector of experiment names of the same length as 
expColors.}

\item{expColors}{character vector of background colors to group the result 
columns belonging to different experiments.}
}
\value{
No value returned.
}
\description{
Produce Excel table of TPP-TR or TPP-CCR experiment out of the
data frame returned by \code{\link{tpptrAnalyzeMeltingCurves}}
}
\examples{
data(hdacTR_resultsTable_smallExample)
tppExport(resultTable, "tpptr_example_results.xlsx")

}
