% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpptrPlotSplines.R
\name{tpptrPlotSplines}
\alias{tpptrPlotSplines}
\title{Plot spline fits per protein}
\usage{
tpptrPlotSplines(data, factorsH1, factorsH0 = c(), fittedModels, testResults,
  resultPath, ggplotTheme = tppDefaultTheme(), plotAlphabetical = FALSE,
  plotRanked = TRUE, maxRank = 500)
}
\arguments{
\item{data}{the data to be plotted.}

\item{factorsH1}{which factors were included in the alternative model?
(necessary for correct prediction)}

\item{factorsH0}{which factors were included in the null model? 
(necessary for correct prediction)}

\item{fittedModels}{long table of fitted models. 
Output of \code{\link{tpptrFitSplines}}.}

\item{testResults}{long table of p-values per protein. 
Output of \code{\link{tpptrFTest}}.}

\item{resultPath}{location where to store the spline plots per protein.}

\item{ggplotTheme}{ggplot theme for melting curve plots.}

\item{plotAlphabetical}{Generate a summary pdf with 20 plots per page in 
alphabetical order?}

\item{plotRanked}{Generate a summary pdf with 20 plots per page, ordered by p-value?}

\item{maxRank}{if \code{plotRanked = TRUE}, how many of the top hits should 
be plotted (default: 500)?}
}
\value{
None
}
\description{
Plot spline fits per protein
}
\details{
Plots of the natural spline fits will be stored in a subfolder with 
name \code{Spline_Fits} at the location specified by \code{resultPath}.
}
\examples{
data(hdacTR_smallExample)
tpptrData <- tpptrImport(configTable = hdacTR_config, data = hdacTR_data)
normResults <- tpptrNormalize(data = tpptrData, normReqs = tpptrDefaultNormReqs())
normData_eSets <- normResults$normData
normData_longTable <- tpptrTidyUpESets(normData_eSets)$proteinMeasurements
splineFits <- tpptrFitSplines(data = normData_longTable,
                              factorsH1 = "condition", returnModels = TRUE)
testResults <- tpptrFTest(fittedModels = splineFits, doPlot = FALSE)
tpptrPlotSplines(data = normData_longTable, fittedModels = splineFits,
                 factorsH1 = "condition", factorsH0 = c(),
                 testResults = testResults, resultPath = getwd(),
                 plotRanked = TRUE, maxRank = 20)

}
\seealso{
\code{\link{ns}, \link{AICc}, \link{tppDefaultTheme}, 
\link{tpptrFitSplines}, \link{tpptrFTest}}
}

