% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertESetsToLongTables.R
\name{tpptrTidyUpESets}
\alias{tpptrTidyUpESets}
\title{Tidy up expressionSets}
\usage{
tpptrTidyUpESets(tppESetList)
}
\arguments{
\item{tppESetList}{A list of expressionSets, which currently still is the format
for intermediate results of most TR and CCR functions.}
}
\value{
A list with two entries: \code{proteinMeasurements} contains the fold 
changes per protein, across all experiments. \code{proteinAnnotation} 
contains additional annotation per protein and experiment. For example, the
peptide counts per identified protein would go here.
}
\description{
Convert list of expressionSets (intermediate output of several 
TR and CCR functions) to list of tidy tables.
}
\details{
expressionSet lists are for example produced by 
\code{\link{tpptrImport}}, \code{\link{tpptrNormalize}}, 
\code{\link{tpptrCurveFit}}, \code{\link{tppccrImport}}, 
\code{\link{tppccrNormalize}}, \code{\link{tppccrCurveFit}}.
}
\examples{
data(hdacTR_smallExample)
tpptrData <- tpptrImport(configTable = hdacTR_config, data = hdacTR_data)
longTables <- tpptrData \%>\% tpptrTidyUpESets
concentrations <- longTables \%>\% extract2("proteinMeasurements")
additionalInfos <- longTables \%>\% extract2("proteinAnnotation")
summary(concentrations)
}

